+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [ 0.4 ] - 07/01/2020
+
+- Added setup.py and uploaded to pip
+- Updated to usage as package
+- Updated print to logging
+- Many small fixes
+
+
 ## [ 0.3 ] - 06/11/2019
 
 - Refactored all functions